refactor Chart component to extract range and limit drag logic into createRangeDragBehavior and createLimitDragBehavior factory functions in useChartDrag module, remove getXandYLocationForChartDrag import, and consolidate drag event handling, state management, and cleanup logic into reusable drag behavior creators
